Added code completion for Editor wrapping types
