+++ src/com/intellij/idea/plugin/hybris/system/cockpitng/codeInsight/lookup/CngLookupElementFactory.kt
@@ -1,6 +1,6 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -18,7 +18,11 @@
 
 package com.intellij.idea.plugin.hybris.system.cockpitng.codeInsight.lookup
 
+import com.intellij.codeInsight.completion.InsertionContext
+import com.intellij.codeInsight.completion.PrioritizedLookupElement
+import com.intellij.codeInsight.lookup.LookupElement
 import com.intellij.codeInsight.lookup.LookupElementBuilder
+import com.intellij.idea.plugin.hybris.codeInsight.completion.AutoPopupInsertHandler
 import com.intellij.idea.plugin.hybris.common.utils.HybrisIcons
 import com.intellij.idea.plugin.hybris.system.cockpitng.meta.model.*
 import com.intellij.psi.xml.XmlTag
@@ -58,4 +62,20 @@ object CngLookupElementFactory {
     fun buildInitializeProperty(property: String) = LookupElementBuilder.create(property)
         .withIcon(HybrisIcons.COCKPIT_NG_INITIALIZE_PROPERTY)
 
+    fun buildWrappingType(lookupString: String, presentableText: String, tailText: String? = null) = PrioritizedLookupElement.withGrouping(
+        PrioritizedLookupElement.withPriority(
+            LookupElementBuilder.create(lookupString)
+                .withPresentableText(presentableText)
+                .withTailText(tailText?.let { " ($it)" }, true)
+                .withTypeText(lookupString, true)
+                .withIcon(HybrisIcons.TYPE_OBJECT)
+                .withInsertHandler(
+                    object : AutoPopupInsertHandler() {
+                        override fun handle(context: InsertionContext, item: LookupElement) {
+                            val cursorOffset = context.editor.caretModel.offset
+                            context.editor.caretModel.moveToOffset(cursorOffset - 1)
+                        }
+                    }
+                ), 2.0
+        ), 1)
 }
